Bug 799451 - Cannot translate different meanings of string…  "_Balance" to Chinese.
The previous fix for this ticket added context for some uses of string
"Balance" but that's not part of the original ticket and doesn't have
ambiguity (it's always used as a noun). The real problem "_Balance" (the
underscore indicates it is a label of an action) however is not fully
covered. This ticket adds context all uses of "_Balance" and removed the
unnecesary contexts for "Balance".