Allow more account types in the payment window's transfer account list
So far it was limited to Asset, Liability, Credit Card
and either Income (for customer payments) or Expenses (for vendor payments).

There are however use cases where a customer payment also covers
expenses (like for example a Paypal service charge).

To allow for as many edge cases as possible the
filter now only excludes those account types I worry
will break the business logic: AP and AR accounts.
That is, with this change you won't be able to create a
customer payment (which necessarily has one split an an AR account)
that transfers to another AR or to an AP account.

While even that may work, I am sure the business features weren't
designed to cope with this and I can't exhaustively test that.